+++ ReBloom/Assets/AddressableAssetsData/AddressableAssetSettings.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   m_DefaultGroup: 85e3454c1ef6fb140be90dbc4071ccf8
   m_currentHash:
     serializedVersion: 2
-    Hash: b7bbc1ea7a3bbe06573c157af444ed2b
+    Hash: 53f062e4073049c91bee022ac91571d2
   m_OptimizeCatalogSize: 0
   m_BuildRemoteCatalog: 0
   m_CatalogRequestsTimeout: 0
