+++ ReBloom/Assets/Arts/Prefabs/Inventory/EmptySlot.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 3266076599690602102}
   - component: {fileID: 4451125446398907089}
   - component: {fileID: 2148394506443669809}
+  - component: {fileID: 2486030122877075528}
   m_Layer: 5
   m_Name: EmptySlot
   m_TagString: Untagged
@@ -77,6 +78,25 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!114 &2486030122877075528
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 165018241709572705}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b72a0911137e1eb4e9c6d8b1325ff67b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  zoneType: 1
+  slotIndex: 0
+  priority: 50
+  background: {fileID: 2148394506443669809}
+  normalColor: {r: 1, g: 1, b: 1, a: 1}
+  hoverColor: {r: 0.8, g: 1, b: 0.8, a: 1}
+  invalidColor: {r: 1, g: 0.8, b: 0.8, a: 1}
 --- !u!1 &5542638675436596870
 GameObject:
   m_ObjectHideFlags: 0
