+++ ReBloom/Assets/Scripts/Build_Storage/UI/StorageUI.cs
@@ -1,18 +1,2 @@
-﻿using System.Collections.Generic; using UnityEngine;  public class StorageUI : MonoBehaviour {     [Header("References")]     [SerializeField] private InventoryItemData playerInventory;     [SerializeField] private Transform slotParent;     [SerializeField] private GameObject emptySlotPrefab;     [SerializeField] private GameObject storageSlotPrefab;     [SerializeField] private StorageDropZone dropZone;      [Header("Settings")]     [SerializeField] private int maxSlots = 20;      private StorageData storageData;     private WorldStorage worldStorage;      private List<Transform> emptySlots = new List<Transform>();     private List<StorageSlot> activeSlots = new List<StorageSlot>();      private bool isInitialized = false;      /// <summary>     /// WorldStorage에서 호출하여 데이터 초기화     /// </summary>     public void Initialize(StorageData data, WorldStorage storage)     {         // 이미 구독되어 있으면 먼저 해제         if (isInitialized && storageData != null)         {             storageData.OnContainerChanged -= RefreshUI;         }          storageData = data;         worldStorage = storage;          // Initialize 시점에 바로 이벤트 구독         if (storageData != null)         {             storageData.OnContainerChanged += RefreshUI;             isInitialized = true;             Debug.Log($"[StorageUI] 초기화 완료 - 이벤트 구독 완료");         }          // DropZone 초기화         if (dropZone != null)         {             dropZone.Initialize(storage);         }     }      private void Awake()     {         CreateEmptySlots();     }      private void OnEnable()     {         // UI 활성화 시 즉시 갱신 (이벤트는 이미 Initialize에서 구독됨)         if (storageData != null)         {             Debug.Log("[StorageUI] OnEnable - UI 갱신 시작");             RefreshUI();         }         else         {             Debug.LogWarning("[StorageUI] OnEnable - StorageData가 아직 초기화되지 않았습니다!");         }     }      private void OnDisable()     {         Debug.Log("[StorageUI] OnDisable");         // OnDisable에서는 이벤트 해제하지 않음 (계속 유지)     }      private void OnDestroy()     {         // 파괴될 때만 이벤트 해제         if (storageData != null)         {             storageData.OnContainerChanged -= RefreshUI;             Debug.Log("[StorageUI] OnDestroy - 이벤트 구독 해제");         }     }      private void CreateEmptySlots()     {         emptySlots.Clear();          for (int i = 0; i < maxSlots; i++)         {             var emptySlot = Instantiate(emptySlotPrefab, slotParent);             emptySlot.SetActive(true);             emptySlots.Add(emptySlot.transform);         }          Debug.Log($"[StorageUI] {maxSlots}개 빈 슬롯 생성 완료");     }      private void RefreshUI()     {         if (storageData == null)         {             Debug.LogWarning("[StorageUI] RefreshUI - StorageData가 없습니다!");             return;         }          Debug.Log("[StorageUI] ===== RefreshUI 시작! =====");          ClearAllItemSlots();          var items = storageData.GetAllItems();         Debug.Log($"[StorageUI] 창고 데이터 아이템 개수: {items.Count}");          for (int i = 0; i < items.Count && i < emptySlots.Count; i++)         {             var itemStack = items[i];             Debug.Log($"[StorageUI] 인덱스 {i}: itemID={itemStack.itemID}, count={itemStack.count}");              if (itemStack.itemID != 0)             {                 var itemBase = ItemDatabase.I.GetItem(itemStack.itemID);                 if (itemBase != null)                 {                     CreateItemSlot(i, itemBase, itemStack.count);                     Debug.Log($"[StorageUI] 슬롯 {i} 생성 완료: {itemBase.itemName} x{itemStack.count}");                 }                 else                 {                     Debug.LogError($"[StorageUI] x ItemDatabase에서 itemID {itemStack.itemID}를 찾을 수 없습니다!");                 }             }         }          Debug.Log($"[StorageUI] ===== UI 갱신 완료: {activeSlots.Count}개 활성 슬롯 =====");     }      private void CreateItemSlot(int slotIndex, ItemBase itemData, int count)     {         if (slotIndex < 0 || slotIndex >= emptySlots.Count)         {             Debug.LogWarning($"[StorageUI] 잘못된 슬롯 인덱스: {slotIndex}");             return;         }          Debug.Log($"[StorageUI] CreateItemSlot 시작 - 슬롯: {slotIndex}, 아이템: {itemData.itemName}");          var slotObj = Instantiate(storageSlotPrefab, emptySlots[slotIndex]);         var storageSlot = slotObj.GetComponent<StorageSlot>();          if (storageSlot != null)         {             storageSlot.SetItem(itemData, count);             activeSlots.Add(storageSlot);             Debug.Log($"[StorageUI] StorageSlot 생성 성공!");         }         else         {             Debug.LogError("[StorageUI] x StorageSlot 컴포넌트를 찾을 수 없습니다!");         }     }      private void ClearAllItemSlots()     {         Debug.Log($"[StorageUI] ClearAllItemSlots - 제거할 슬롯: {activeSlots.Count}개");          foreach (var slot in activeSlots)         {             if (slot != null)             {                 Destroy(slot.gameObject);             }         }         activeSlots.Clear();     }      public void WithdrawItem(int slotIndex)     {         if (storageData == null || playerInventory == null)         {             Debug.LogError("[StorageUI] 데이터가 없습니다!");             return;         }          var items = storageData.GetAllItems();         if (slotIndex < 0 || slotIndex >= items.Count)             return;          var itemStack = items[slotIndex];         if (itemStack.itemID == 0)             return;          bool added = playerInventory.AddItem(itemStack.itemID, itemStack.count);         if (added)         {             storageData.RemoveItem(itemStack.itemID, itemStack.count);             Debug.Log($"[StorageUI] {itemStack.count}개 회수 완료");         }     }      public void Close()     {         gameObject.SetActive(false);
-        Cursor.visible = false;
-        Cursor.lockState = CursorLockMode.Locked;     }
-
-    public void Toggle()
-    {
-        gameObject.SetActive(!gameObject.activeSelf);
-        if (gameObject.activeSelf)
-        {
-            Cursor.visible = true;
-            Cursor.lockState = CursorLockMode.None;
-        }
-        else
-        {
-            Cursor.visible = false;
-            Cursor.lockState = CursorLockMode.Locked;
-        }
-    } }
+﻿using System.Collections.Generic; using UnityEngine;  public class StorageUI : MonoBehaviour {     [Header("References")]     [SerializeField] private InventoryItemData playerInventory;     [SerializeField] private Transform slotParent;     [SerializeField] private GameObject emptySlotPrefab;     [SerializeField] private GameObject storageSlotPrefab;      [Header("Settings")]     [SerializeField] private int maxSlots = 20;      private StorageData storageData;     private WorldStorage worldStorage;      private List<Transform> emptySlots = new List<Transform>();     private List<StorageSlot> activeSlots = new List<StorageSlot>();      private bool isInitialized = false;      /// <summary>     /// WorldStorage에서 호출하여 데이터 초기화     /// </summary>     public void Initialize(StorageData data, WorldStorage storage)     {         // 이미 구독되어 있으면 먼저 해제         if (isInitialized && storageData != null)         {             storageData.OnContainerChanged -= RefreshUI;         }          storageData = data;         worldStorage = storage;          // Initialize 시점에 바로 이벤트 구독         if (storageData != null)         {             storageData.OnContainerChanged += RefreshUI;             isInitialized = true;             Debug.Log($"[StorageUI] 초기화 완료 - 이벤트 구독 완료");         }          // DropZone 초기화         if (DragDropManager.I != null)         {
+            DragDropManager.I.SetCurrentStorage(storage);         }     }      private void Awake()     {         CreateEmptySlots();     }      private void OnEnable()     {         // UI 활성화 시 즉시 갱신 (이벤트는 이미 Initialize에서 구독됨)         if (storageData != null)         {             Debug.Log("[StorageUI] OnEnable - UI 갱신 시작");             RefreshUI();         }         else         {             Debug.LogWarning("[StorageUI] OnEnable - StorageData가 아직 초기화되지 않았습니다!");         }     }      private void OnDisable()     {         Debug.Log("[StorageUI] OnDisable");         // OnDisable에서는 이벤트 해제하지 않음 (계속 유지)     }      private void OnDestroy()     {         // 파괴될 때만 이벤트 해제         if (storageData != null)         {             storageData.OnContainerChanged -= RefreshUI;             Debug.Log("[StorageUI] OnDestroy - 이벤트 구독 해제");         }     }      private void CreateEmptySlots()     {         emptySlots.Clear();          for (int i = 0; i < maxSlots; i++)         {             var emptySlot = Instantiate(emptySlotPrefab, slotParent);             emptySlot.SetActive(true);             emptySlots.Add(emptySlot.transform);         }          Debug.Log($"[StorageUI] {maxSlots}개 빈 슬롯 생성 완료");     }      private void RefreshUI()     {         if (storageData == null)         {             Debug.LogWarning("[StorageUI] RefreshUI - StorageData가 없습니다!");             return;         }          Debug.Log("[StorageUI] ===== RefreshUI 시작! =====");          ClearAllItemSlots();          var items = storageData.GetAllItems();         Debug.Log($"[StorageUI] 창고 데이터 아이템 개수: {items.Count}");          for (int i = 0; i < items.Count && i < emptySlots.Count; i++)         {             var itemStack = items[i];             Debug.Log($"[StorageUI] 인덱스 {i}: itemID={itemStack.itemID}, count={itemStack.count}");              if (itemStack.itemID != 0)             {                 var itemBase = ItemDatabase.I.GetItem(itemStack.itemID);                 if (itemBase != null)                 {                     CreateItemSlot(i, itemBase, itemStack.count);                     Debug.Log($"[StorageUI] 슬롯 {i} 생성 완료: {itemBase.itemName} x{itemStack.count}");                 }                 else                 {                     Debug.LogError($"[StorageUI] x ItemDatabase에서 itemID {itemStack.itemID}를 찾을 수 없습니다!");                 }             }         }          Debug.Log($"[StorageUI] ===== UI 갱신 완료: {activeSlots.Count}개 활성 슬롯 =====");     }      private void CreateItemSlot(int slotIndex, ItemBase itemData, int count)     {         if (slotIndex < 0 || slotIndex >= emptySlots.Count)         {             Debug.LogWarning($"[StorageUI] 잘못된 슬롯 인덱스: {slotIndex}");             return;         }          Debug.Log($"[StorageUI] CreateItemSlot 시작 - 슬롯: {slotIndex}, 아이템: {itemData.itemName}");          var slotObj = Instantiate(storageSlotPrefab, emptySlots[slotIndex]);         var storageSlot = slotObj.GetComponent<StorageSlot>();          if (storageSlot != null)         {             storageSlot.SetItem(itemData, count);             activeSlots.Add(storageSlot);             Debug.Log($"[StorageUI] StorageSlot 생성 성공!");         }         else         {             Debug.LogError("[StorageUI] x StorageSlot 컴포넌트를 찾을 수 없습니다!");         }     }      private void ClearAllItemSlots()     {         Debug.Log($"[StorageUI] ClearAllItemSlots - 제거할 슬롯: {activeSlots.Count}개");          foreach (var slot in activeSlots)         {             if (slot != null)             {                 Destroy(slot.gameObject);             }         }         activeSlots.Clear();     }      public void WithdrawItem(int slotIndex)     {         if (storageData == null || playerInventory == null)         {             Debug.LogError("[StorageUI] 데이터가 없습니다!");             return;         }          var items = storageData.GetAllItems();         if (slotIndex < 0 || slotIndex >= items.Count)             return;          var itemStack = items[slotIndex];         if (itemStack.itemID == 0)             return;          bool added = playerInventory.AddItem(itemStack.itemID, itemStack.count);         if (added)         {             storageData.RemoveItem(itemStack.itemID, itemStack.count);             Debug.Log($"[StorageUI] {itemStack.count}개 회수 완료");         }     }      public void Close()     {         gameObject.SetActive(false);         Cursor.visible = false;         Cursor.lockState = CursorLockMode.Locked;     }      public void Toggle()     {         gameObject.SetActive(!gameObject.activeSelf);         if (gameObject.activeSelf)         {             Cursor.visible = true;             Cursor.lockState = CursorLockMode.None;         }         else         {             Cursor.visible = false;             Cursor.lockState = CursorLockMode.Locked;         }     } }
