--- ReBloom/Assets/Scripts/DragDrop/DragContext.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 52f6bc9e513e11340adeb2ae75c66167
