--- ReBloom/Assets/Scripts/DragDrop/IDragSource.cs
@@ -1,1 +1,0 @@
-﻿/// <summary> /// 드래그 출발지 인터페이스 /// </summary> public interface IDragSource {     DragSourceType SourceType { get; }     int SlotIndex { get; }      /// <summary>     /// 드래그 시작 시 호출     /// </summary>     DragContext CreateDragContext(ItemBase item);      /// <summary>     /// 드롭 성공 후 출발지 처리 (아이템 제거 등)     /// </summary>     void OnDragSuccess();      /// <summary>     /// 드롭 실패 시 원복     /// </summary>     void OnDragCancelled(); }
