--- ReBloom/Assets/Scripts/DragDrop/IDragSource.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4ad85aa6bc4abf245acbe76bfe825e64
