--- ReBloom/Assets/Scripts/DragDrop/IDropSource.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 34d34366a87103442860c5ae151e7303
