--- ReBloom/Assets/Scripts/DragDrop/ItemIconDragHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 217ae252225d99c4e9acf8d4e39a980e
