--- ReBloom/Assets/Scripts/DragDrop/QuickSlotDropZone.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8047a50f008ba0549b3350be6b5f8ed6
