+++ ReBloom/Assets/Scripts/DragDrop/Refactor/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 442490519ad690242886bc48bd623184
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
