--- ReBloom/Assets/Scripts/DragDrop/Refactor/DragDropEnum.cs
@@ -1,24 +1,0 @@
-﻿/// <summary>
-/// 드래그 출발지 타입
-/// </summary>
-public enum DragSourceType
-{
-    Inventory,
-    QuickSlot,
-    Equipment,
-    Storage,
-    Debug
-}
-
-/// <summary>
-/// 드롭존 타입
-/// </summary>
-public enum DropZoneType
-{
-    QuickSlot,
-    Inventory,
-    Storage,
-    World,
-    Equipment,
-    TrashBin
-}
