--- ReBloom/Assets/Scripts/DragDrop/Refactor/DragDropEnum.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fd9751128b7f060489741e83b6d401cb
