--- ReBloom/Assets/Scripts/DragDrop/Refactor/DragDropManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 892243ee63078134f898098efe5cf047
