--- ReBloom/Assets/Scripts/DragDrop/Refactor/DropZoneMarker.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b72a0911137e1eb4e9c6d8b1325ff67b
