+++ ReBloom/Assets/Scripts/DragDrop/Refactor/Handler/ItemIconDragHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 217ae252225d99c4e9acf8d4e39a980e
