+++ ReBloom/Assets/Scripts/DragDrop/Refactor/Manager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d39864d9e67f9dd44a951264ffb64e5a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
