+++ ReBloom/Assets/Scripts/DragDrop/Refactor/Manager/DragDropManager.cs
@@ -1,0 +1,39 @@
+﻿using System.Collections.Generic; using Cysharp.Threading.Tasks; using UnityEngine; using UnityEngine.EventSystems;  /// <summary> /// 드래그앤드롭 중앙 관리자 (Marker 기반) /// 모든 드래그 상태와 로직을 한 곳에서 관리 /// </summary> public class DragDropManager : MonoBehaviour {     public static DragDropManager I { get; private set; }      [Header("System References")]     [SerializeField] private QuickSlot quickSlotManager;     [SerializeField] private InventoryItemData inventoryData;     [SerializeField] private ItemSpawner itemSpawner;     [SerializeField] private RemovePopUp removePopUp;     [SerializeField] private ItemDropCalculator itemDropCalculator;      [Header("Debug Settings")]     [SerializeField] private int debugSpawnCount = 1;     [SerializeField] private bool showDebugLogs = true;      // 현재 드래그 상태     private DragContext currentDrag;     private WorldStorage currentStorage;     private DropZoneMarker currentHoveredZone;      #region 이벤트 시스템     /// <summary>     /// 드래그 중 전역 피드백 이벤트     /// (예: 퀵슬롯 불가능 아이템 경고)     /// </summary>     public static event System.Action<DragContext, bool> OnDragFeedback;     #endregion      #region Singleton     private void Awake()     {         if (I != null && I != this)         {             Destroy(gameObject);             return;         }         I = this;          // ItemDropCalculator 자동 찾기         if (itemDropCalculator == null)         {             itemDropCalculator = FindFirstObjectByType<ItemDropCalculator>();             if (itemDropCalculator == null)             {                 Debug.LogWarning("[DragDropManager] ItemDrop을 찾을 수 없습니다! 월드 드롭이 작동하지 않을 수 있습니다.");             }         }
+         //Debug.Log("[DragDropManager] 초기화 완료");     }     #endregion      #region Public API     /// <summary>     /// 드래그 시작     /// </summary>     public void BeginDrag(DragContext context)     {         if (context == null || context.Item == null)         {             Debug.LogWarning("[DragDropManager] 유효하지 않은 컨텍스트입니다.");             return;         }          currentDrag = context;         OnDragFeedback?.Invoke(context, true);          Debug.Log($"드래그 시작: {context.Item.itemName} from {context.SourceType}");     }      /// <summary>     /// 드래그 종료 및 드롭 처리     /// </summary>     public void EndDrag(PointerEventData eventData)     {         if (currentDrag == null)         {             Debug.LogWarning("[DragDropManager] 드래그 컨텍스트가 없습니다.");             return;         }          // RaycastAll로 모든 UI 요소 검색         List<RaycastResult> results = new List<RaycastResult>();         EventSystem.current.RaycastAll(eventData, results);          // 최적 드롭존 찾기 (우선순위 기반)         DropZoneMarker bestZone = FindBestDropZone(results);          // 드롭 처리         if (bestZone != null && CanDrop(bestZone, currentDrag))         {             HandleDrop(bestZone, currentDrag);             currentDrag.Source?.OnDragSuccess();            // Debug.Log($"드롭 성공: {bestZone.ZoneType} (슬롯: {bestZone.SlotIndex})");         }         else         {             currentDrag.Source?.OnDragCancelled();           //  Debug.Log("드롭 실패 - 원위치 복원");         }          OnDragFeedback?.Invoke(null, false);          // 비주얼 초기화         if (currentHoveredZone != null)         {             currentHoveredZone.ResetVisual();             currentHoveredZone = null;         }          currentDrag = null;     }      /// <summary>     /// 호버 처리 (비주얼 피드백)     /// </summary>     public void NotifyHover(PointerEventData eventData)     {         if (currentDrag == null) return;          // RaycastAll로 검색         List<RaycastResult> results = new List<RaycastResult>();         EventSystem.current.RaycastAll(eventData, results);          DropZoneMarker hoveredZone = FindBestDropZone(results);          // 이전 호버 초기화         if (currentHoveredZone != hoveredZone)         {             currentHoveredZone?.ResetVisual();             currentHoveredZone = hoveredZone;         }          // 드롭 가능 여부 판단         bool canDrop = hoveredZone != null && CanDrop(hoveredZone, currentDrag);          // 전역 피드백 이벤트 발생         OnDragFeedback?.Invoke(currentDrag, canDrop);          // 새 호버 표시         if (hoveredZone != null)         {             if (CanDrop(hoveredZone, currentDrag))             {                 hoveredZone.ShowValidHover();             }             else             {                 hoveredZone.ShowInvalidHover();             }         }     }      /// <summary>     /// 현재 창고 설정 (StorageUI에서 호출)     /// </summary>     public void SetCurrentStorage(WorldStorage storage)     {         currentStorage = storage;         Debug.Log($"창고 설정: {storage?.name}");     }      /// <summary>     /// 팝업에서 호출 - 확정된 수량으로 월드 드롭     /// </summary>     public async void DropItemFromPopup(ItemBase item, int quantity)     {         if (item == null || quantity <= 0)         {             Debug.LogWarning("[DragDropManager] 유효하지 않은 아이템 또는 수량입니다.");             return;         }          Vector3 dropPosition = GetDropPosition();         await DropToWorldAsync(item, dropPosition, quantity);     }     #endregion      #region Private Methods - 드롭존 선택 및 검증     /// <summary>     /// 우선순위 기반 최적 드롭존 선택     /// </summary>     private DropZoneMarker FindBestDropZone(List<RaycastResult> results)     {         DropZoneMarker bestZone = null;         int highestPriority = -1;          foreach (var result in results)         {             var zone = result.gameObject.GetComponent<DropZoneMarker>();             if (zone != null && zone.Priority > highestPriority)             {                 highestPriority = zone.Priority;                 bestZone = zone;             }         }          return bestZone;     }      /// <summary>     /// 드롭 가능 여부 검증     /// </summary>     private bool CanDrop(DropZoneMarker zone, DragContext context)     {         switch (zone.ZoneType)         {             case DropZoneType.QuickSlot:                 return CanDropToQuickSlot(zone, context);              case DropZoneType.Storage:                 return context.IsFromInventory && currentStorage != null;              case DropZoneType.World:                 return true;              case DropZoneType.Inventory:                 return true; // TODO: 인벤토리 검증 로직              default:                 return false;         }     }      private bool CanDropToQuickSlot(DropZoneMarker zone, DragContext context)     {         // canQuickSlot 플래그 체크         if (!context.Item.canQuickSlot) return false;          // 같은 슬롯 체크         if (context.IsFromQuickSlot && context.SourceSlotIndex == zone.SlotIndex) return false;          return true;     }     #endregion      #region Private Methods - 드롭 처리     /// <summary>     /// 실제 드롭 처리 (타입별 분기)     /// </summary>     private void HandleDrop(DropZoneMarker zone, DragContext context)     {         switch (zone.ZoneType)         {             case DropZoneType.QuickSlot:                 HandleQuickSlotDrop(zone.SlotIndex, context);                 break;              case DropZoneType.Storage:                 HandleStorageDrop(context);                 break;              case DropZoneType.World:                 HandleWorldDrop(context);                 break;              case DropZoneType.Inventory:                 HandleInventoryDrop(zone.SlotIndex, context);                 break;         }     }      /// <summary>     /// 퀵슬롯 드롭 처리     /// </summary>     private void HandleQuickSlotDrop(int targetSlot, DragContext context)     {         if (context.IsFromQuickSlot)         {             // 퀵슬롯 간 스왑             SwapQuickSlots(context.SourceSlotIndex, targetSlot);         }         else         {             // 인벤토리 -> 퀵슬롯             AssignToQuickSlot(targetSlot, context.Item);         }     }      /// <summary>     /// 창고 드롭 처리     /// </summary>     private void HandleStorageDrop(DragContext context)     {         if (currentStorage == null)         {             Debug.LogError("[DragDropManager] 창고가 설정되지 않았습니다!");             return;         }          int quantity = inventoryData.GetItemCount(context.Item.itemID);          if (inventoryData.RemoveItem(context.Item.itemID, quantity))         {             currentStorage.AddItem(context.Item, quantity);             Debug.Log($"창고에 {context.Item.itemName} x{quantity} 보관 완료");         }     }      /// <summary>     /// 월드 드롭 처리     /// </summary>     private void HandleWorldDrop(DragContext context)     {         if (context.IsFromDebug)         {             // 디버그: 즉시 드롭             Vector3 dropPosition = GetDropPosition();             DropToWorldAsync(context.Item, dropPosition, debugSpawnCount).Forget();         }         else if (context.IsFromQuickSlot)         {             // 퀵슬롯에서 제거만 (인벤토리는 유지)             quickSlotManager?.RemoveSlot(context.SourceSlotIndex);            // Debug.Log($"퀵슬롯 {context.SourceSlotIndex}에서 제거 (인벤토리 유지)");         }         else         {             // 인벤토리: 수량 선택 팝업             removePopUp?.OnOpen(context.Item);         }     }
+
+    /// <summary>     /// 인벤토리 드롭 처리     /// </summary>         private void HandleInventoryDrop(int targetSlot, DragContext context)
+    {
+        if (inventoryData == null)
+        {
+            Debug.LogError("[DragDropManager] InventoryItemData를 찾을 수 없습니다!");
+            return;
+        }
+
+        // 같은 슬롯에 드롭: 무시
+        if (context.IsFromInventory && context.SourceSlotIndex == targetSlot)
+        {
+            Debug.Log($"[DragDropManager] 인벤토리 같은 슬롯 드롭 무시: {targetSlot}");
+            return;
+        }
+
+        // 인벤토리 -> 인벤토리: 스왑
+        if (context.IsFromInventory)
+        {
+            // InventoryItemData의 SwapSlots 메서드 사용
+            bool success = inventoryData.SwapSlots(context.SourceSlotIndex, targetSlot);
+
+            if (success)
+            {
+                Debug.Log($"[DragDropManager] 인벤토리 스왑 완료: {context.SourceSlotIndex} <-> {targetSlot}");
+            }
+            else
+            {
+                Debug.LogWarning($"[DragDropManager] 인벤토리 스왑 실패: {context.SourceSlotIndex} <-> {targetSlot}");
+            }
+        }
+        else
+        {
+            // TODO: 다른 소스 -> 인벤토리 (예: 퀵슬롯 -> 인벤토리)
+            Debug.Log($"[DragDropManager] 인벤토리 드롭: 슬롯 {targetSlot} (다른 소스는 미구현)");
+        }
+    }     #endregion      #region Helper Methods - 퀵슬롯     /// <summary>     /// 퀵슬롯 간 스왑     /// </summary>     private void SwapQuickSlots(int fromIndex, int toIndex)     {         if (quickSlotManager == null)         {             Debug.LogError("[DragDropManager] QuickSlot 매니저가 없습니다!");             return;         }          ItemBase fromItem = quickSlotManager.GetItemAtSlot(fromIndex);         ItemBase toItem = quickSlotManager.GetItemAtSlot(toIndex);          GameInventory gameInventory = FindFirstObjectByType<GameInventory>();         if (gameInventory == null)         {             Debug.LogError("[DragDropManager] GameInventory를 찾을 수 없습니다!");             return;         }          int fromQuantity = gameInventory.GetItemCount(fromItem.itemID);         int toQuantity = toItem != null ? gameInventory.GetItemCount(toItem.itemID) : 0;         // Debug.Log($"퀵슬롯 스왑: [{fromIndex}] {fromItem.itemName} <-> [{toIndex}] {(toItem != null ? toItem.itemName : "빈 슬롯")}");          // 1. 두 슬롯 모두 제거         quickSlotManager.RemoveSlot(fromIndex);         if (toItem != null)         {             quickSlotManager.RemoveSlot(toIndex);         }          // 2. 교차 배치         quickSlotManager.AssignToSlot(toIndex, fromItem, fromQuantity);         if (toItem != null)         {             quickSlotManager.AssignToSlot(fromIndex, toItem, toQuantity);         }     }      /// <summary>     /// 인벤토리 → 퀵슬롯 배치     /// </summary>     private void AssignToQuickSlot(int slotIndex, ItemBase item)     {         if (quickSlotManager == null)         {             Debug.LogError("[DragDropManager] QuickSlot 매니저가 없습니다!");             return;         }          GameInventory gameInventory = FindFirstObjectByType<GameInventory>();         int quantity = gameInventory != null ? gameInventory.GetItemCount(item.itemID) : 0;          if (quantity <= 0)         {             Debug.LogWarning($"[DragDropManager] 인벤토리에 {item.itemName}이(가) 없습니다!");             return;         }          // 이미 다른 슬롯에 등록되어 있는지 확인         int existingSlotIndex = quickSlotManager.FindItemSlot(item);         if (existingSlotIndex >= 0 && existingSlotIndex != slotIndex)         {             Debug.Log($"{item.itemName}이(가) 슬롯 {existingSlotIndex}에 있어서 제거 후 슬롯 {slotIndex}에 재배치");             quickSlotManager.RemoveSlot(existingSlotIndex);         }          // 현재 슬롯에 기존 아이템 있으면 교체         ItemBase currentSlotItem = quickSlotManager.GetItemAtSlot(slotIndex);         if (currentSlotItem != null)         {           //  Debug.Log($"슬롯 {slotIndex}: {currentSlotItem.itemName} → {item.itemName} (교체)");             quickSlotManager.RemoveSlot(slotIndex);         }          // 새 아이템 배치         bool success = quickSlotManager.AssignToSlot(slotIndex, item, quantity);         if (success)         {            // Debug.Log($"{item.itemName} x{quantity}를 슬롯 {slotIndex}에 배치 완료");         }     }     #endregion      #region Helper Methods - 월드 드롭     /// <summary>     /// 드롭 위치 가져오기 (ItemDropCalculator 사용)     /// </summary>     private Vector3 GetDropPosition()     {         if (itemDropCalculator != null)         {             return itemDropCalculator.CalculateDropPosition();         }          // Fallback: 플레이어 앞         GameObject player = GameObject.FindGameObjectWithTag("Player");         if (player != null)         {             return player.transform.position + player.transform.forward * 2f + Vector3.up * 1.5f;         }          Debug.LogError("[DragDropManager] 드롭 위치를 계산할 수 없습니다!");         return Vector3.zero;     }      /// <summary>     /// 월드에 아이템 드롭 (async)     /// </summary>     private async UniTask DropToWorldAsync(ItemBase item, Vector3 dropPosition, int quantity)     {         if (itemSpawner == null)         {             Debug.LogError("[DragDropManager] ItemSpawner를 찾을 수 없습니다!");             return;         }          // 아이템 스폰         await itemSpawner.DropItemWithQuantity(item, dropPosition, quantity);         InventroyEventSystem.ItemDropped();         // Debug.Log($"{item.itemName} x{quantity}개를 월드에 드롭했습니다.");     }     #endregion }
