+++ ReBloom/Assets/Scripts/DragDrop/Refactor/Manager/DragDropManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 892243ee63078134f898098efe5cf047
