+++ ReBloom/Assets/Scripts/DragDrop/Refactor/Manager/ItemDropCalculator.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0b189a2535931e842b641a73344c9f26
