--- ReBloom/Assets/Scripts/DragDrop/StorageDropZone.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7771e6f79abbaca4b97ee5793a38ba69
