--- ReBloom/Assets/Scripts/DragDrop/WorldDropZone.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dea7c1932ca17ab46bfa58ab787fe52f
